Move common functionality into Core module
