+++ Core/src/main/java/com/arivanamin/healthcare/backend/base/domain/audit/AuditEventPublisher.java
@@ -1,0 +1,7 @@
+package com.arivanamin.healthcare.backend.base.domain.audit;
+
+@FunctionalInterface
+public interface AuditEventPublisher {
+    
+    void sendAuditLog (String topic, AuditEvent event);
+}
