+++ Patient-Service/src/main/java/com/arivanamin/healthcare/backend/patient/application/endpoints/PatientController.java
@@ -1,11 +1,9 @@
 package com.arivanamin.healthcare.backend.patient.application.endpoints;
 
-import com.arivanamin.healthcare.backend.base.domain.audit.AuditEvent;
-import com.arivanamin.healthcare.backend.base.domain.audit.AuditTopics;
+import com.arivanamin.healthcare.backend.base.domain.audit.*;
 import com.arivanamin.healthcare.backend.patient.application.request.CreatePatientRequest;
 import com.arivanamin.healthcare.backend.patient.application.request.UpdatePatientRequest;
 import com.arivanamin.healthcare.backend.patient.application.response.*;
-import com.arivanamin.healthcare.backend.patient.core.audit.AuditEventPublisher;
 import com.arivanamin.healthcare.backend.patient.core.command.*;
 import com.arivanamin.healthcare.backend.patient.core.query.ReadPatientByIdQuery;
 import com.arivanamin.healthcare.backend.patient.core.query.ReadPatientsQuery;
@@ -18,7 +16,6 @@ import org.springframework.cache.annotation.Cacheable;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
-import java.time.LocalDateTime;
 import java.util.UUID;
 
 import static com.arivanamin.healthcare.backend.patient.application.config.PatientApiConfig.PROTECTED_API_BASE_PATH;
@@ -36,6 +33,7 @@ class PatientController {
     private final DeletePatientCommand deleteCommand;
     
     private final AuditEventPublisher auditPublisher;
+    private final AuditEvent event;
     
     @GetMapping (PROTECTED_API_BASE_PATH + "/v1/accounts")
     @Cacheable (cacheNames = "patientsCache")
@@ -50,13 +48,11 @@ class PatientController {
     @Operation (summary = "Get a single patient by id")
     @ResponseStatus (HttpStatus.OK)
     public PatientResponse getPatientById (@PathVariable UUID id) {
-        AuditEvent event = AuditEvent.builder()
-            .location(PROTECTED_API_BASE_PATH + "/v1/accounts/{id})")
-            .data(String.valueOf(id))
-            .serviceName("transaction-service")
-            .timestamp(LocalDateTime.now())
-            .action("get patient by id")
-            .build();
+        
+        event.setLocation(PROTECTED_API_BASE_PATH + "/v1/accounts/{id})");
+        event.setData(String.valueOf(id));
+        event.setAction("get patient by id");
+        
         auditPublisher.sendAuditLog(AuditTopics.API_AUDIT_TOPIC, event);
         return PatientResponse.of(readByIdQuery.execute(id));
     }
